add get subscriber list to admin resources

+++ src/main/java/de/ipk_gatersleben/bit/bi/bridge/brapicomp/resources/AdminResource.java
@@ -3,9 +3,11 @@ package de.ipk_gatersleben.bit.bi.bridge.brapicomp.resources;
 import java.io.IOException;
 import java.io.InputStream;
 import java.sql.SQLException;
+import java.util.List;
 
 import javax.ws.rs.Consumes;
 import javax.ws.rs.DELETE;
+import javax.ws.rs.GET;
 import javax.ws.rs.POST;
 import javax.ws.rs.Path;
 import javax.ws.rs.Produces;
@@ -24,6 +26,7 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 
 import de.ipk_gatersleben.bit.bi.bridge.brapicomp.Config;
 import de.ipk_gatersleben.bit.bi.bridge.brapicomp.dbentities.Endpoint;
+import de.ipk_gatersleben.bit.bi.bridge.brapicomp.dbentities.EndpointService;
 import de.ipk_gatersleben.bit.bi.bridge.brapicomp.dbentities.TestReport;
 import de.ipk_gatersleben.bit.bi.bridge.brapicomp.testing.config.TestCollection;
 import de.ipk_gatersleben.bit.bi.bridge.brapicomp.utils.DataSourceManager;
@@ -42,11 +45,13 @@ public class AdminResource {
 
     private static final Logger LOGGER = LogManager.getLogger(AdminResource.class.getName());
 
-    /**
+    //Commented out for security. Also private API key is required
+    /*
+    *//**
      * Run the default test on all endpoints
      *
      * @return Response with json report
-     */
+     *//*
     @POST
     @Path("/testall")
     @Consumes(MediaType.APPLICATION_JSON)
@@ -89,10 +94,24 @@ public class AdminResource {
 
     @DELETE
     @Path("/tables")
-    public Response delTables() {
+    public Response delTables(@Context HttpHeaders headers) {
     	// It is necessary to delete or update the database tables after the schema changes
     	// i.e. when a new parameter/column is added
         try {
+        	
+        	String[] auth = ResourceService.getAuth(headers);
+            //Check auth header
+            if (auth == null || auth.length != 2) {
+                String e = JsonMessageManager.jsonMessage(401, "unauthorized", 4000);
+                return Response.status(Status.UNAUTHORIZED).entity(e).build();
+            }
+
+            //Check if api key is correct.
+            if (!auth[1].equals(Config.get("adminkey"))) {
+                String e = JsonMessageManager.jsonMessage(403, "missing or wrong apikey", 4001);
+                return Response.status(Status.UNAUTHORIZED).entity(e).build();
+            }
+            
             DataSourceManager.deleteTable(TestReport.class);
             DataSourceManager.deleteTable(Endpoint.class);
         } catch (SQLException e) {
@@ -100,4 +119,32 @@ public class AdminResource {
         }
         return Response.ok().build();
     }
+    @GET
+    @Path("/users")
+    public Response getUsers(@Context HttpHeaders headers) {
+    	// It is necessary to delete or update the database tables after the schema changes
+    	// i.e. when a new parameter/column is added
+    	List<Endpoint> l;
+        try {
+        	
+        	String[] auth = ResourceService.getAuth(headers);
+            //Check auth header
+            if (auth == null || auth.length != 2) {
+                String e = JsonMessageManager.jsonMessage(401, "unauthorized", 4000);
+                return Response.status(Status.UNAUTHORIZED).entity(e).build();
+            }
+
+            //Check if api key is correct.
+            if (!auth[1].equals(Config.get("adminkey"))) {
+                String e = JsonMessageManager.jsonMessage(403, "missing or wrong apikey", 4001);
+                return Response.status(Status.UNAUTHORIZED).entity(e).build();
+            }
+        	
+            l = EndpointService.getAllEndpoints();
+            
+        } catch (SQLException e) {
+            return Response.serverError().entity(e.getMessage()).build();
+        }
+        return Response.ok().entity(l.toString()).build();
+    }*/
 }
